+++ libs/wire-api/src/Wire/API/Routes/Public/Galley.hs
@@ -204,9 +204,10 @@ data Api routes = Api
     getConversationByReusableCode ::
       routes
         :- Summary "Get limited conversation information by key/code pair"
-        :> CanThrow NotConnected
-        :> CanThrow OperationDenied
         :> CanThrow NotATeamMember
+        :> CanThrow CodeNotFound
+        :> CanThrow ConvNotFound
+        :> CanThrow ConvAccessDenied
         :> ZUser
         :> "conversations"
         :> "join"
@@ -216,10 +217,9 @@ data Api routes = Api
     createGroupConversation ::
       routes
         :- Summary "Create a new conversation"
+        :> CanThrow NotConnected
+        :> CanThrow OperationDenied
         :> CanThrow NotATeamMember
-        :> CanThrow CodeNotFound
-        :> CanThrow ConvNotFound
-        :> CanThrow ConvAccessDenied
         :> Description "This returns 201 when a new conversation is created, and 200 when the conversation already existed"
         :> ZUser
         :> ZConn
@@ -370,7 +370,37 @@ data Api routes = Api
         :- FeatureStatusGet 'TeamFeatureFileSharing,
     teamFeatureStatusFileSharingPut ::
       routes
-        :- FeatureStatusPut 'TeamFeatureFileSharing
+        :- FeatureStatusPut 'TeamFeatureFileSharing,
+    teamFeatureStatusClassifiedDomainsGet ::
+      routes
+        :- FeatureStatusGet 'TeamFeatureClassifiedDomains,
+    featureAllFeatureConfigsGet ::
+      routes
+        :- AllFeatureConfigsGet,
+    featureConfigLegalHoldGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureLegalHold,
+    featureConfigSSOGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureSSO,
+    featureConfigSearchVisibilityGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureSearchVisibility,
+    featureConfigValidateSAMLEmailsGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureValidateSAMLEmails,
+    featureConfigDigitalSignaturesGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureDigitalSignatures,
+    featureConfigAppLockGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureAppLock,
+    featureConfigFileSharingGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureFileSharing,
+    featureConfigClassifiedDomainsGet ::
+      routes
+        :- FeatureConfigGet 'TeamFeatureClassifiedDomains
   }
   deriving (Generic)
 
@@ -415,6 +445,19 @@ type FeatureStatusDeprecatedPut featureName =
     :> DeprecatedFeatureName featureName
     :> ReqBody '[Servant.JSON] (TeamFeatureStatus featureName)
     :> Put '[Servant.JSON] (TeamFeatureStatus featureName)
+
+type FeatureConfigGet featureName =
+  Summary (AppendSymbol "Get feature config for feature " (KnownTeamFeatureNameSymbol featureName))
+    :> ZUser
+    :> "feature-configs"
+    :> KnownTeamFeatureNameSymbol featureName
+    :> Get '[Servant.JSON] (TeamFeatureStatus featureName)
+
+type AllFeatureConfigsGet =
+  Summary "Get configurations of all features"
+    :> ZUser
+    :> "feature-configs"
+    :> Get '[Servant.JSON] AllFeatureConfigs
 
 type PostOtrDescriptionUnqualified =
   "This endpoint ensures that the list of clients is correct and only sends the message if the list is correct.\n\