+++ libs/wire-api/src/Wire/API/Team/Feature.hs
@@ -22,6 +22,7 @@ module Wire.API.Team.Feature
   ( TeamFeatureName (..),
     TeamFeatureStatus,
     TeamFeatureAppLockConfig (..),
+    TeamFeatureClassifiedDomainsConfig (..),
     TeamFeatureStatusValue (..),
     FeatureHasNoConfig,
     EnforceAppLock (..),
@@ -29,7 +30,9 @@ module Wire.API.Team.Feature
     TeamFeatureStatusNoConfig (..),
     TeamFeatureStatusWithConfig (..),
     HasDeprecatedFeatureName (..),
+    AllFeatureConfigs (..),
     defaultAppLockStatus,
+    defaultClassifiedDomains,
 
     -- * Swagger
     typeTeamFeatureName,
@@ -37,12 +40,16 @@ module Wire.API.Team.Feature
     modelTeamFeatureStatusNoConfig,
     modelTeamFeatureStatusWithConfig,
     modelTeamFeatureAppLockConfig,
+    modelTeamFeatureClassifiedDomainsConfig,
     modelForTeamFeature,
   )
 where
 
+import Control.Lens.Combinators (dimap)
+import qualified Data.Aeson as Aeson
 import qualified Data.Attoparsec.ByteString as Parser
 import Data.ByteString.Conversion (FromByteString (..), ToByteString (..), toByteString')
+import Data.Domain (Domain)
 import Data.Kind (Constraint)
 import Data.Schema
 import Data.String.Conversions (cs)
@@ -59,15 +66,27 @@ import Wire.API.Arbitrary (Arbitrary, GenericUniform (..))
 ----------------------------------------------------------------------
 -- TeamFeatureName
 
--- | If you add a constructor here, you need to visit (at least) 4 places that are not caught
--- by ghc errors or test failures:
+-- | If you add a constructor here, you need extend multiple defintions, which
+--   aren't checked by GHC.
 --
--- * libs/wire-api/test/unit/Test/Wire/API/Roundtrip/Aeson.hs (calls to 'testRoundTrip')
--- * services/galley/src/Galley/API/Internal.hs (add a field to the 'InternalApi routes' record)
--- * libs/wire-api/src/Wire/API/Routes/Public/Galley.hs (add a field to the 'Api routes' record)
--- * services/galley/src/Galley/API/Teams/Features.hs (calls to 'getStatus')
--- * services/galley/schema/src/ (add a migration like the one in "V43_TeamFeatureDigitalSignatures.hs")
--- * services/galley/test/integration/API/Teams/Feature.hs (add an integration test)
+--   Follow this Checklist:
+--
+-- * libs/wire-api/test/unit/Test/Wire/API/Roundtrip/Aeson.hs
+--   * add call to 'testRoundTrip'
+-- * services/galley/src/Galley/API/Internal.hs
+--   * add a field to the 'InternalApi routes' record)
+-- * libs/wire-api/src/Wire/API/Routes/Public/Galley.hs
+--   * add a GET (and possible PUT) route with name prefix teamFeature<FEATURE_NAME>
+--   * add a GET route with name prefix featureConfig<FEATURE_NAME>
+-- * services/galley/src/Galley/API/Teams/Features.hs
+--   * extend getAllFeatureConfigs
+--   * extend getAllFeatures
+-- * services/galley/schema/src/
+--   * add a migration like the one in "V43_TeamFeatureDigitalSignatures.hs"
+-- * services/galley/test/integration/API/Teams/Feature.hs
+--   * add an integration test for the feature
+--   * extend testAllFeatures
+--
 --
 -- An overview of places to change (including compiler errors and failing tests) can be found
 -- in eg. https://github.com/wireapp/wire-server/pull/1652.
@@ -94,6 +113,7 @@ data TeamFeatureName
   | TeamFeatureDigitalSignatures
   | TeamFeatureAppLock
   | TeamFeatureFileSharing
+  | TeamFeatureClassifiedDomains
   deriving stock (Eq, Show, Ord, Generic, Enum, Bounded, Typeable)
   deriving (Arbitrary) via (GenericUniform TeamFeatureName)
 
@@ -129,6 +149,10 @@ instance KnownTeamFeatureName 'TeamFeatureFileSharing where
   type KnownTeamFeatureNameSymbol 'TeamFeatureFileSharing = "fileSharing"
   knownTeamFeatureName = TeamFeatureFileSharing
 
+instance KnownTeamFeatureName 'TeamFeatureClassifiedDomains where
+  type KnownTeamFeatureNameSymbol 'TeamFeatureClassifiedDomains = "classifiedDomains"
+  knownTeamFeatureName = TeamFeatureClassifiedDomains
+
 instance FromByteString TeamFeatureName where
   parser =
     Parser.takeByteString >>= \b ->
@@ -144,8 +168,11 @@ instance FromByteString TeamFeatureName where
         Right "digital-signatures" -> pure TeamFeatureDigitalSignatures
         Right "appLock" -> pure TeamFeatureAppLock
         Right "fileSharing" -> pure TeamFeatureFileSharing
+        Right "classifiedDomains" -> pure TeamFeatureClassifiedDomains
         Right t -> fail $ "Invalid TeamFeatureName: " <> T.unpack t
 
+-- TODO: how do we make this consistent with 'KnownTeamFeatureNameSymbol'?  add a test for
+-- that?  anyway do we really need both?
 instance ToByteString TeamFeatureName where
   builder TeamFeatureLegalHold = "legalhold"
   builder TeamFeatureSSO = "sso"
@@ -154,6 +181,14 @@ instance ToByteString TeamFeatureName where
   builder TeamFeatureDigitalSignatures = "digitalSignatures"
   builder TeamFeatureAppLock = "appLock"
   builder TeamFeatureFileSharing = "fileSharing"
+  builder TeamFeatureClassifiedDomains = "classifiedDomains"
+
+instance ToSchema TeamFeatureName where
+  schema =
+    enum @Text
+      "TeamFeatureName"
+      $ mconcat
+        (map (\feat -> element (cs . toByteString' $ feat) feat) [minBound .. maxBound])
 
 class HasDeprecatedFeatureName (a :: TeamFeatureName) where
   type DeprecatedFeatureName a :: Symbol
@@ -220,6 +255,7 @@ type family TeamFeatureStatus (a :: TeamFeatureName) :: * where
   TeamFeatureStatus 'TeamFeatureDigitalSignatures = TeamFeatureStatusNoConfig
   TeamFeatureStatus 'TeamFeatureAppLock = TeamFeatureStatusWithConfig TeamFeatureAppLockConfig
   TeamFeatureStatus 'TeamFeatureFileSharing = TeamFeatureStatusNoConfig
+  TeamFeatureStatus 'TeamFeatureClassifiedDomains = TeamFeatureStatusWithConfig TeamFeatureClassifiedDomainsConfig
 
 type FeatureHasNoConfig (a :: TeamFeatureName) = (TeamFeatureStatus a ~ TeamFeatureStatusNoConfig) :: Constraint
 
@@ -232,6 +268,7 @@ modelForTeamFeature TeamFeatureValidateSAMLEmails = modelTeamFeatureStatusNoConf
 modelForTeamFeature TeamFeatureDigitalSignatures = modelTeamFeatureStatusNoConfig
 modelForTeamFeature name@TeamFeatureAppLock = modelTeamFeatureStatusWithConfig name modelTeamFeatureAppLockConfig
 modelForTeamFeature TeamFeatureFileSharing = modelTeamFeatureStatusNoConfig
+modelForTeamFeature name@TeamFeatureClassifiedDomains = modelTeamFeatureStatusWithConfig name modelTeamFeatureClassifiedDomainsConfig
 
 ----------------------------------------------------------------------
 -- TeamFeatureStatusNoConfig
@@ -284,6 +321,31 @@ instance ToSchema cfg => ToSchema (TeamFeatureStatusWithConfig cfg) where
         <*> tfwcConfig .= field "config" schema
 
 ----------------------------------------------------------------------
+-- TeamFeatureClassifiedDomainsConfig
+
+newtype TeamFeatureClassifiedDomainsConfig = TeamFeatureClassifiedDomainsConfig
+  { classifiedDomainsDomains :: [Domain]
+  }
+  deriving stock (Show, Eq, Generic)
+  deriving (ToJSON, FromJSON, S.ToSchema) via (Schema TeamFeatureClassifiedDomainsConfig)
+
+deriving via (GenericUniform TeamFeatureClassifiedDomainsConfig) instance Arbitrary TeamFeatureClassifiedDomainsConfig
+
+instance ToSchema TeamFeatureClassifiedDomainsConfig where
+  schema =
+    object "TeamFeatureClassifiedDomainsConfig" $
+      TeamFeatureClassifiedDomainsConfig
+        <$> classifiedDomainsDomains .= field "domains" (array schema)
+
+modelTeamFeatureClassifiedDomainsConfig :: Doc.Model
+modelTeamFeatureClassifiedDomainsConfig =
+  Doc.defineModel "TeamFeatureClassifiedDomainsConfig" $ do
+    Doc.property "domains" (Doc.array Doc.string') $ Doc.description "domains"
+
+defaultClassifiedDomains :: TeamFeatureStatusWithConfig TeamFeatureClassifiedDomainsConfig
+defaultClassifiedDomains = TeamFeatureStatusWithConfig TeamFeatureDisabled (TeamFeatureClassifiedDomainsConfig [])
+
+----------------------------------------------------------------------
 -- TeamFeatureAppLockConfig
 
 data TeamFeatureAppLockConfig = TeamFeatureAppLockConfig
@@ -330,3 +392,12 @@ data LowerCaseFirst
 instance StringModifier LowerCaseFirst where
   getStringModifier (x : xs) = toLower x : xs
   getStringModifier [] = []
+
+newtype AllFeatureConfigs = AllFeatureConfigs {_allFeatureConfigs :: Aeson.Object}
+  deriving stock (Eq, Show)
+  deriving (FromJSON, ToJSON, S.ToSchema) via (Schema AllFeatureConfigs)
+
+instance ToSchema AllFeatureConfigs where
+  schema =
+    named "AllFeatureConfigs" $
+      dimap _allFeatureConfigs AllFeatureConfigs jsonObject