+++ CHANGELOG.md
@@ -24,21 +24,42 @@
 
 ## Release Notes
 
+If you want to set the default for file sharing in all teams to `disabled`, add the following to `galley.yaml` (default is "enabled"):
+
+```
+settings:
+  featureFlags:
+    fileSharing:
+      defaults:
+        status: disabled
+```
+
+## Release Notes for Wire.com Cloud operators
+
+Upgrade nginz (#1658)
+
 ## API Changes
 
 ## Features
 
-* `fileSharing` feature config (#1652)
+* Extend feature config API (#1658)
+* `fileSharing` feature config (#1652, #1654, #1655)
 * Add user_id to csv export (#1663)
 
 ## Bug fixes and other updates
 
 * New, hardened end-point for changing email
-* Remove old end-point for changing email
+* Fix: CSV export is missing SCIM external id when SAML is also used (#1608)
+* Fix: sso_id field in user record (brig) was not always filled correctly in cassandra (#1334)
 
 ## Documentation
 
+* Improved Swagger documentation for endpoints with multiple responses (#1649, #1645).
+
 ## Internal changes
+
+* The servant-swagger dependency now points to the current upstream master (#1656).
+* Refactor function createUser for readability (#1670)
 
 ## Federation changes (alpha feature, do not use yet)
 
@@ -51,6 +72,9 @@ This release requires a manual change in your galley configuration: `settings.co
 
 ## API Changes
 
+* A new team feature for classified domains is available (#1626):
+  - a public endpoint is at `GET /teams/:tid/features/classifiedDomains`
+  - an internal endpoint is at `GET /i/teams/:tid/features/classifiedDomains`
 * Several public team feature endpoints are removed (their internal and
   Stern-based counterparts remain available):
   - `PUT /teams/:tid/features/sso`
@@ -92,7 +116,6 @@ This release requires a manual change in your galley configuration: `settings.co
 ## Documentation
 
 * Fix validation errors in Swagger documentation (#1625).
-* Improved Swagger documentation for endpoints with multiple responses (#1649, #1645).
 
 ## Bug fixes and other updates